Rename login classes
