+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,10 @@
         android:icon="${icon}"
         android:isGame="true"
         android:label="@string/app_name"
-        android:usesCleartextTraffic="true"
         android:roundIcon="${roundIcon}"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Pluvia">
+        android:theme="@style/Theme.Pluvia"
+        android:usesCleartextTraffic="true">
         <!--
          android:windowSoftInputMode must always be in the activity block.
          IME padding with composables like 'Scaffold' will have terrible resizing consequences, if in the activity tag.
@@ -31,11 +31,10 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:windowSoftInputMode="adjustResize"
-            android:theme="@style/Theme.Pluvia">
+            android:theme="@style/Theme.Pluvia"
+            android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
             <intent-filter>
